==> processes/02_data-formatting/03_convert-census-csvs-to-list-of-tibbles_2022-10-25.qmd <==
---
title: "Reading Census Files"
author: "Ladun"
format: html
editor: visual
---

### Libraries

```{r libs}
library(here)
library(vroom)
library(fs)
library(purrr)
library(stringr)
```

## Read csvs

### Create read_census_csvs function

```{r tableRead}
read_census_csvs <- function(path){
  
  path_csvs <- dir(path, full.names = TRUE)
  
  tables <- lapply(path_csvs, vroom, show_col_types = FALSE) # additional arg to avoid huge visual clutter
  table_names <-
    str_extract(path_csvs, "[^/]+(?=$)") |>
    str_remove(".csv") |> 
    str_extract("[^-]+(?=$)")

  names(tables) <- table_names
  
  
  return(tables)
}

rename_list <- function(x, paths){
  new_names <- str_extract(paths, "[^/]+(?=$)")
  
  names(x) <- new_names
  
  return(x)
}

```

### Read CoM csvs

```{r readCoMCsvs}
com_path <- here("data", "inputs", "census", "com", "csvs")
com_dirs <- dir_ls(com_path)

com_tables <-
  map(com_dirs, read_census_csvs) |>
  rename_list(com_dirs)
```

### Read Nation csvs

```{r readNationcsvs}
nation_path <- here("data", "inputs", "census", "nations", "csvs")
nation_dirs <- dir_ls(nation_path)

nation_tables <-
  map(nation_dirs, read_census_csvs) |> 
  rename_list(nation_dirs)
```

### Read Diocese csvs

```{r readDiocesecsvs}
dioceses_path <- here("data", "inputs", "census", "dioceses", "csvs")
dioceses_dirs <- dir_ls(dioceses_path)

dioceses_tables <-
  map(dioceses_dirs, read_census_csvs) |> 
  rename_list(dioceses_dirs)
```

## Make big list and save

```{r saveList}

overall_table <-
  list(
    com = com_tables,
    nation = nation_tables,
    dioceses = dioceses_tables
    )

saveRDS(
  overall_table,
  file = here("data", "intermediate", "census", "merged-data",
              "01_all-census-data_list_2022-10-12.RDS")
  )
```
